Making float/fixed more compatible, Tilemap support

+++ sdl2/sdl2main.cpp
@@ -48,8 +48,8 @@ static void redraw() {
 }
 
 void setBacklight(f32 value) {
-    value = (value < 0) ? 0 : value;
-    value = (value > 1) ? 1 : value;
+    value = (value < 0) ? f32(0) : value;
+    value = (value > 1) ? f32(1) : value;
     backlight = value;
 }
